# Desenvolvimento_de_Aplicações_Web_ASP_.NET_MVC_e_Web_API/Livraria/LivrariaAPI/Request/CriarAutorRequest.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace LivrariaAPI.Request
{
    public class CriarAutorRequest
    {
        public string Nome { get; set; }
        public string Sobrenome { get; set; }
        public DateTime Datanascimento { get; set; }
        public string Email { get; set; }
        public string Senha { get; set; }
    }
}